+++ src/main/java/net/more_rpg_classes/MRPGCMod.java
@@ -1,8 +1,11 @@
 package net.more_rpg_classes;
 
 import net.fabricmc.api.ModInitializer;
+import net.minecraft.registry.Registries;
+import net.minecraft.registry.Registry;
 import net.more_rpg_classes.client.particle.MoreParticles;
 import net.more_rpg_classes.config.EffectsConfig;
+import net.more_rpg_classes.custom.Enchantment_CustomSpellSchool;
 import net.more_rpg_classes.custom.MoreSpellSchools;
 import net.more_rpg_classes.effect.MRPGCEffects;
 import net.more_rpg_classes.entity.attribute.MRPGCEntityAttributes;
@@ -14,6 +17,7 @@ import net.tinyconfig.ConfigManager;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import static net.spell_power.SpellPowerMod.enchantmentConfig;
 
 public class MRPGCMod implements ModInitializer {
 	public static final String MOD_ID = "more_rpg_classes";
@@ -38,6 +42,13 @@ public class MRPGCMod implements ModInitializer {
 		ModSounds.register();
 		MoreSpellSchools.initialize();
 
+			for(var entry: Enchantment_CustomSpellSchool.all.entrySet()) {
+				Registry.register(Registries.ENCHANTMENT, entry.getKey(), entry.getValue());
+			}
+			enchantmentConfig.value.apply();
+			Enchantment_CustomSpellSchool.attach();
 	}
+
 }
+