+++ src/main/java/net/more_rpg_classes/client/particle/MoreParticles.java
@@ -11,12 +11,40 @@ public class MoreParticles {
     public static final DefaultParticleType STUN_PAR = FabricParticleTypes.simple();
     public static final DefaultParticleType BLOOD_DROP = FabricParticleTypes.simple();
     public static final DefaultParticleType MOLTEN_ARMOR= FabricParticleTypes.simple();
+    public static final DefaultParticleType BUBBLE = FabricParticleTypes.simple();
+    public static final DefaultParticleType BUBBLE_POP = FabricParticleTypes.simple();
+    public static final DefaultParticleType WATER_MIST = FabricParticleTypes.simple();
+    public static final DefaultParticleType SPLASH = FabricParticleTypes.simple();
+    public static final DefaultParticleType BIG_SPLASH = FabricParticleTypes.simple();
+    public static final DefaultParticleType WAVE = FabricParticleTypes.simple();
+    public static final DefaultParticleType DRIPPING_WATER = FabricParticleTypes.simple();
+    public static final DefaultParticleType HOT_SPLASH = FabricParticleTypes.simple();
+    public static final DefaultParticleType WATER_WHIP = FabricParticleTypes.simple();
+    public static final DefaultParticleType WATER_CIRCLE = FabricParticleTypes.simple();
+    public static final DefaultParticleType WATER_HEAL = FabricParticleTypes.simple();
+    public static final DefaultParticleType WATER_SPLASH = FabricParticleTypes.simple();
+    public static final DefaultParticleType STONE_EXPLOSION = FabricParticleTypes.simple();
+    public static final DefaultParticleType STONE_PARTICLE = FabricParticleTypes.simple();
 
 
     public static void register(){
         Registry.register(Registries.PARTICLE_TYPE, new Identifier(MRPGCMod.MOD_ID, "stun_particle"), STUN_PAR);
         Registry.register(Registries.PARTICLE_TYPE, new Identifier(MRPGCMod.MOD_ID, "blood_drop"), BLOOD_DROP);
         Registry.register(Registries.PARTICLE_TYPE, new Identifier(MRPGCMod.MOD_ID, "molten_armor"), MOLTEN_ARMOR);
+        Registry.register(Registries.PARTICLE_TYPE, new Identifier(MRPGCMod.MOD_ID, "bubble"), BUBBLE);
+        Registry.register(Registries.PARTICLE_TYPE, new Identifier(MRPGCMod.MOD_ID, "bubble_pop"), BUBBLE_POP);
+        Registry.register(Registries.PARTICLE_TYPE, new Identifier(MRPGCMod.MOD_ID, "water_mist"), WATER_MIST);
+        Registry.register(Registries.PARTICLE_TYPE, new Identifier(MRPGCMod.MOD_ID, "splash"), SPLASH);
+        Registry.register(Registries.PARTICLE_TYPE, new Identifier(MRPGCMod.MOD_ID, "big_splash"), BIG_SPLASH);
+        Registry.register(Registries.PARTICLE_TYPE, new Identifier(MRPGCMod.MOD_ID, "wave"), WAVE);
+        Registry.register(Registries.PARTICLE_TYPE, new Identifier(MRPGCMod.MOD_ID, "dripping_water"), DRIPPING_WATER);
+        Registry.register(Registries.PARTICLE_TYPE, new Identifier(MRPGCMod.MOD_ID, "hot_splash"), HOT_SPLASH);
+        Registry.register(Registries.PARTICLE_TYPE, new Identifier(MRPGCMod.MOD_ID, "water_whip"), WATER_WHIP);
+        Registry.register(Registries.PARTICLE_TYPE, new Identifier(MRPGCMod.MOD_ID, "water_circle"), WATER_CIRCLE);
+        Registry.register(Registries.PARTICLE_TYPE, new Identifier(MRPGCMod.MOD_ID, "water_heal"), WATER_HEAL);
+        Registry.register(Registries.PARTICLE_TYPE, new Identifier(MRPGCMod.MOD_ID, "water_splash"), WATER_SPLASH);
+        Registry.register(Registries.PARTICLE_TYPE, new Identifier(MRPGCMod.MOD_ID, "stone_explosion"), STONE_EXPLOSION);
+        Registry.register(Registries.PARTICLE_TYPE, new Identifier(MRPGCMod.MOD_ID, "stone_particle"), STONE_PARTICLE);
     }
 
 }