+++ src/main/java/net/more_rpg_classes/config/EffectsConfig.java
@@ -9,5 +9,5 @@ public class EffectsConfig {
     public float grievous_wounds_healing_taken = 0.25F;
     public float collected_soul_soul_power_per_stack = 0.10F;
     public float frozen_solid_increased_damage_taken = 0.5F;
-    public float frosted_decreased_movement_speed_per_stack = -1.1F;
+    public float frosted_decreased_movement_speed_per_stack = -0.1F;
 }