+++ src/main/java/net/more_rpg_classes/effect/FrostedEffect.java
@@ -1,5 +1,7 @@
 package net.more_rpg_classes.effect;
 
+import net.minecraft.entity.LivingEntity;
+import net.minecraft.entity.attribute.AttributeContainer;
 import net.minecraft.entity.effect.StatusEffect;
 import net.minecraft.entity.effect.StatusEffectCategory;
 
@@ -8,4 +10,8 @@ public class FrostedEffect extends StatusEffect {
         super(statusEffectCategory, color);
     }
 
+    @Override
+    public void onApplied(LivingEntity entity, AttributeContainer attributes, int amplifier) {
+        entity.setFrozenTicks(entity.getFrozenTicks() + 200);
+    }
 }